create a stored procedure that computes and store the average for a student

+++ 0x00-MySQL_Advanced/7-average_score.sql
@@ -6,7 +6,7 @@ DELIMITER $$ ;
 
 CREATE PROCEDURE ComputeAverageScoreForUser(IN user_id INT)
 BEGIN
-	DECLARE average_score INT;
+	DECLARE average_score FLOAT;
 
 	SELECT AVG(score) INTO average_score 
 		FROM corrections
